setu-backend/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
pymongo==4.7.2
pytest==8.2.1
python-dotenv==1.0.1
requests==2.32.2
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.3
